Modified file structure and uploading early-bird challenges
Going to follow YYYY-MM-DD structure so that it'll be easier to ID in the future.

Also, decided to do a challenge in the morning as warm up for the rest of the day.

Daily Coding sends them out in the afternoon... so I checked out leetcode for something to work on